Added testing functionality for storeGenre
So this was a long time coming but I needed to set up a nodeJs environment for testing javascript files outside of the browser. I included detailed comments for the first test implemented (storeGenre.test.js) in the test directory for explanations on why I used what I used. Chai is the assertion library for testing, and sinon is used for extra functionality + sinon-chrome via acvetkov to create a sandbox where a dummy chrome storage API can be used.

Co-Authored-By: A-bmw <108549880+A-bmw@users.noreply.github.com>