UD4155-36 As a user, I would like to have a form with which I can sub…  …mit answers to my preferred genres.
This adds full functionality for the form via corrections to form.js and form.html. in-depth comments included.  It also displays the userVector array and the userResponses array. As well, it currently has a 3 second timer which then sends users to a skeleton genremap.html. Also included are some styling changes that make the entire experience UP TO genremap.html inline with experience within the app so far. More styling should be done, but for now it is solid enough.

Co-Authored-By: Chanc10 <89817601+Chanc10@users.noreply.github.com>
Co-Authored-By: ochsner3k <40705326+ochsner3k@users.noreply.github.com>