UD4155-7 As a reader who is interested in books of all genres, I wish to be able to get/give suggestions based off of the genres I have selected.
this is the beginning of the structure of recommendation scoring. So far, I've been able to call to the books API, and add book data to the console. Currently similarity score isn't working, but my books are displaying. I'm pretty sure cosine similarity works but I'm  putting this up so there's some iterative commit. I also made some minor fixes to content.js, and form.js. index.html got a test link for recommendations while I work. Test.html is a page to click on a button. for recommendations. They print to the console.

Co-Authored-By: Chanc10 <89817601+Chanc10@users.noreply.github.com>
Co-Authored-By: ochsner3k <40705326+ochsner3k@users.noreply.github.com>