UD4155-7 As a reader who is interested in books of all genres, I wish…  … to be able to get/give suggestions based off of the genres I have selected.
fixed the similarity score issue, and properly displayed book data now to the user.
The issue was a misunderstanding of what I was doing with the books array. Instead of displaying in the genreVector, it now displays to the user separately in the recommendations object. The recommendations object now grabs books data from the genreVector.

I also fixed the recommendations to ONLY display book data if the user wants the genre to be displayed via similarity score being non-zero.
I also removed some arbitrary comments.

LFG

Co-Authored-By: Chanc10 <89817601+Chanc10@users.noreply.github.com>
Co-Authored-By: ochsner3k <40705326+ochsner3k@users.noreply.github.com>