UD4155-7 As a reader who is interested in books of all genres, I wis…  …h… … to be able to get/give suggestions based off of the genres I have selected.
Okay, so now the API key is properly stored instead of being in the url for await. This is better practice. I also explained the concept of ternary ops on line 34. Just if someone else reads this.

added a catch for recommendation errors, I was having issues with recommendation.books not properly defining itself.

changed the recommendations.forEach so that it would now properly display the top 40 books instead of the top book.

In hindsight, I think I did similarity score wrong -- instead of calculating for some nuance, I simply used 0 and 1 for no interest and interest respectively. I may adjust this. This is why as of right now, genreVector's values are generally arbitrary.

Co-Authored-By: Chanc10 <89817601+Chanc10@users.noreply.github.com>
Co-Authored-By: ochsner3k <40705326+ochsner3k@users.noreply.github.com>